refactor: divided tests into separate files, added tests for buy_card
